P9/Services/chatbot-service/requirements.txt
flask
flask_cors
python-dotenv
requests
langchain
langchain-openai
